backend/requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PyTorch and deep learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Image processing and computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Data science and machine learning
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Kaggle API integration
kaggle>=1.5.16

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.4.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0